Moved to plastic, coming back for portfolio
Since I started working with a few classmates I decided to try out a new form of source control, Plastic SCM.

I have been using that since, we are already at the 48 update in there, but after having a talk with one of my lecturers about showing my code and progress, I have decided to start updating both in Plastic and Git to make the updates accessible to my portfolio once the game is done instead of just the final piece.

A lot of things have changed since the initial commit, including some problem solving from here to there, one that can be seen is in the loading bar which was recently divided and the normalization had to be fixed to work with values under 0.

From now on the regular updates comments will be more normal, this was just an update as to why I didnt update the git until now.