Layer Sorting & Pathfinding
Redid the layers to be able to change their collisions in the Layer Collision Matrix more easily. Feel like it may need to be done again once more in the future, but for now is working.

Started working on Pathfinding using A* Pathfind algorithm. For this I have decided to follow a tutorial since this would be my first time dealing with pathfinding. If possible it will be edited once completed to include the possibility of destroying environment on the costs. Heap optimization has already been implemented.

Added new comments, and renamed a few public variables.